requirements.txt
arrow==0.13.1
astroid==2.2.5
binaryornot==0.4.4
certifi==2019.3.9
chardet==3.0.4
Click==7.0
dj-database-url==0.5.0
Django==2.1.7
django-environ==0.4.5
django-heroku==0.3.1
django-rest-auth==0.9.3
djangorestframework==3.9.2
djangorestframework-jwt==1.11.0
entrypoints==0.3
flake8==3.7.7
future==0.17.1
gunicorn==19.9.0
idna==2.8
isort==4.3.15
Jinja2==2.10
jinja2-time==0.2.0
lazy-object-proxy==1.3.1
MarkupSafe==1.1.1
mccabe==0.6.1
Pillow==5.4.1
poyo==0.4.2
psycopg2==2.7.7
pycodestyle==2.5.0
pyflakes==2.1.1
PyJWT==1.7.1
pylint==2.3.1
python-dateutil==2.8.0
pytz==2018.9
requests==2.21.0
six==1.12.0
typed-ast==1.3.1
urllib3==1.24.1
whichcraft==0.5.2
whitenoise==4.1.2
wrapt==1.11.1
